Add catches for common bugs
The calculator now can take multiple operators.
For example, if 12 + 3 is entered and then the
- button is clicked. The display will show 15.
Then the user can click a new set of numbers,
i.e. 29 and the equals button and the calculator
will calculate and display the answer to 15-29.
This feature is made in a way so that it can loop
infinitely. So therefore, the user can input: an
operation like 12+4, followed by subsequent operators
and numbers infinitely until equals is pressed.

The calculator can now longer take in subsequent
operators. like 12 + +, if a second operator is
detected before a new number, nothing will happen.

Now the calculator can easily begin a new
calculation by pressing a new number after an
expression is evaluated.

There is a bug where if an operator is used after
an expression is evaluated with equals, the
calculator breaks