Add bug fixes, Add backspace
Now the calculator cannot take in multiple decimal
points.

The calculator now has a working backspace button
(clear). This currently works in evaluating the
first number, the operator and after an evaluation
has been calculated (acts as AC button in this case)
however it currently has a bug associated with
trying to backspace the number entered after an
operator.